dailybuild: trivial warnings の抑止（src/Samples)
git-svn-id: http://springhead.info/spr2/Springhead2/trunk@6791 05cee5c3-a2e9-0310-9523-9dfc2f93dbe1

+++ src/Samples/Foundation/UTPreciseTimerSample/main.cpp
@@ -97,7 +97,7 @@ int _cdecl main(int argc, char* argv[]){
 
 	// 計測データをcsvで出力
 	CSVOUT << "count" << "," << "Precise timer [s]" << "," << "QPTimer [s]" << std::endl;
-	for(int i = 0; i < time.size(); i++){
+	for(size_t i = 0; i < time.size(); i++){
 		CSVOUT << i << "," << time[i].x << "," << time[i].y << std::endl;
 	}
